sync components/annotations (extracting from components/operations.py)
